remove pokemon navigation component
