--- src/app/shared/components/pokedex-select-dropdown/pokedex-select-dropdown.component.html
@@ -1,8 +1,0 @@
-<dd-dropdown
-  [sections]="sections"
-  [placeholder]="
-    this.currentVersionGroupName
-      ? this.currentVersionGroupName
-      : 'Select Pokédex'
-  "
-></dd-dropdown>
